pyproject.toml
[tool.poetry]
name = "tsgpt"
version = "0.1.0"
description = ""
authors = ["FloweryScythe13 <brichett13@gmail.com>"]
readme = "README.md"
packages = [{include = "scr"}]

[tool.poetry.dependencies]
python = "^3.9"
bentoml = "1.3.0"
langchain = "0.2.13"
openai = "^0.27.8"
huggingface-hub = "^0.16.4"
pandas = "^2.0.3"
datasets = "^2.14.4"
sf-hamilton = "^1.26.5"
openpyxl = "^3.1.2"
python-decouple = "^3.8"
faiss-cpu = "^1.8.0.post1"
tiktoken = "^0.7.0"
langchain-community = "^0.2.12"
neo4j = "^5.23.1"
fastapi = "^0.112.2"
opentelemetry-sdk = "1.20.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
